requirements.txt
english-word==1.1.0
